Move single use components into event details screen
