+++ src/screens/EventsScreen/Dialog.test.js
