+++ src/screens/EventsScreen/FilterHeader.test.js
