+++ src/screens/EventsScreen/FilterHeaderCategories.test.js
