+++ src/screens/EventsScreen/FilterHeaderConnected.test.js
